requirements.txt
gdal >= 2
pyyaml
ruamel.yaml
future
pandas
numpy
netcdf4
networkx >= 2
msgpack
xarray